Restructure newGame / youDie to avoid circular dependencies
